requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.18.0
Werkzeug==2.0.1
gunicorn==20.1.0
checkov==2.0.431
pyyaml==5.4.1
